feat: revamp core documentation service
